@@ -7,8 +7,8 @@ plugins {
     id 'jacoco'
 }
 
-sourceCompatibility = '1.8'
-targetCompatibility = '1.8'
+sourceCompatibility = '11'
+targetCompatibility = '11'
 
 intellij {
     version = ideaVersion //for a full list of IntelliJ IDEA releases please see https://www.jetbrains.com/intellij-repository/releases
@@ -53,13 +53,13 @@ configurations {
     }
     compileKotlin {
         kotlinOptions {
-            jvmTarget = "1.8"
+            jvmTarget = "11"
         }
     }
 
     compileTestKotlin {
         kotlinOptions {
-            jvmTarget = "1.8"
+            jvmTarget = "11"
         }
     }
     lsp
@@ -145,20 +145,20 @@ dependencies {
     implementation("io.quarkus:quarkus-arc:$quarkusVersion") {
         transitive = false
     }
-    implementation 'org.eclipse.lsp4mp:org.eclipse.lsp4mp.ls:0.5.0'
+    implementation 'org.eclipse.lsp4mp:org.eclipse.lsp4mp.ls:0.6.0'
     implementation 'org.eclipse.lsp4j:org.eclipse.lsp4j:0.15.0'
     //required by lsp4j as the version from IJ is incompatible
     implementation 'com.google.code.gson:gson:2.8.9'
     implementation 'com.vladsch.flexmark:flexmark:0.50.50'
     testImplementation 'com.redhat.devtools.intellij:intellij-common-ui-test-library:0.2.0'
-    lsp('org.eclipse.lsp4mp:org.eclipse.lsp4mp.ls:0.5.0:uber') {
+    lsp('org.eclipse.lsp4mp:org.eclipse.lsp4mp.ls:0.6.0:uber') {
         transitive = false
     }
-    lsp ('com.redhat.microprofile:com.redhat.quarkus.ls:0.12.1') {
+    lsp ('com.redhat.microprofile:com.redhat.quarkus.ls:0.13.0') {
         transitive = false
     }
-    implementation 'com.redhat.microprofile:com.redhat.qute.ls:0.12.1'
-    lsp ('com.redhat.microprofile:com.redhat.qute.ls:0.12.1:uber') {
+    implementation 'com.redhat.microprofile:com.redhat.qute.ls:0.13.0'
+    lsp ('com.redhat.microprofile:com.redhat.qute.ls:0.13.0:uber') {
         transitive = false
     }
     implementation files(new File(buildDir, 'server')) {
