@@ -13,6 +13,7 @@ import com.intellij.openapi.vfs.VirtualFile;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 /**
  * Configuration file API
@@ -25,7 +26,9 @@ public interface IConfigSource {
 
 	/**
 	 * Returns the property from the given <code>key</code> and null otherwise.
-	 * 
+	 *
+	 * Any property expressions in the property value should not be expanded.
+	 *
 	 * @param key the key
 	 * @return the property from the given <code>key</code> and null otherwise.
 	 */
@@ -34,7 +37,9 @@ public interface IConfigSource {
 	/**
 	 * Returns the property as Integer from the given <code>key</code> and null
 	 * otherwise.
-	 * 
+	 *
+	 * Any property expressions in the property value should not be expanded.
+	 *
 	 * @param key the key
 	 * @return the property as Integer from the given <code>key</code> and null
 	 *         otherwise.
@@ -87,4 +92,12 @@ public interface IConfigSource {
 	 * @return the profile of the config source and null otherwise.
 	 */
 	String getProfile();
+
+	/**
+	 * Returns the keys of all values defined in this config source
+	 *
+	 * @return the keys of all values defined in this config source
+	 */
+	Set<String> getAllKeys();
+
 }
