@@ -9,13 +9,26 @@
 *******************************************************************************/
 package com.redhat.devtools.intellij.lsp4mp4ij.psi.core.project;
 
+import com.intellij.ProjectTopics;
+import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.components.Service;
 import com.intellij.openapi.components.ServiceManager;
 import com.intellij.openapi.module.Module;
+import com.intellij.openapi.project.ModuleListener;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
+import com.intellij.openapi.vfs.VirtualFileManager;
+import com.intellij.openapi.vfs.newvfs.BulkFileListener;
+import com.intellij.openapi.vfs.newvfs.events.VFileContentChangeEvent;
+import com.intellij.openapi.vfs.newvfs.events.VFileCreateEvent;
+import com.intellij.openapi.vfs.newvfs.events.VFileDeleteEvent;
+import com.intellij.openapi.vfs.newvfs.events.VFileEvent;
+import com.intellij.util.messages.MessageBusConnection;
+import com.redhat.devtools.intellij.lsp4mp4ij.psi.internal.core.ls.PsiUtilsLSImpl;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -32,10 +45,46 @@ public final class PsiMicroProfileProjectManager {
 		return ServiceManager.getService(project, PsiMicroProfileProjectManager.class);
 	}
 
-	private final Map<Module, PsiMicroProfileProject> projects;
+	private Project project;
 
-	private PsiMicroProfileProjectManager() {
+	private final Map<Module, PsiMicroProfileProject> projects;
+	private MicroProfileProjectListener microprofileProjectListener;
+
+	private class MicroProfileProjectListener implements ModuleListener, BulkFileListener {
+		@Override
+		public void after(@NotNull List<? extends VFileEvent> events) {
+			for(VFileEvent event : events) {
+				if ((event instanceof VFileDeleteEvent || event instanceof VFileContentChangeEvent ||
+						event instanceof VFileCreateEvent) && isConfigSource(event.getFile())) {
+					Module javaProject = PsiUtilsLSImpl.getInstance(project).getModule(event.getFile());
+					if (javaProject != null) {
+						PsiMicroProfileProject mpProject = getJDTMicroProfileProject(javaProject);
+						if (mpProject != null) {
+							mpProject.evictConfigSourcesCache();
+						}
+					}
+
+				}
+			}
+		}
+
+		@Override
+		public void beforeModuleRemoved(@NotNull Project project, @NotNull Module module) {
+			evict(module);
+		}
+
+		private void evict(Module javaProject) {
+			if (javaProject != null) {
+				// Remove the JDTMicroProfile project instance from the cache.
+				projects.remove(javaProject);
+			}
+		}
+	}
+
+	private PsiMicroProfileProjectManager(Project project) {
+		this.project = project;
 		this.projects = new HashMap<>();
+		initialize();
 	}
 
 	public PsiMicroProfileProject getJDTMicroProfileProject(Module project) {
@@ -63,5 +112,15 @@ public final class PsiMicroProfileProjectManager {
 			}
 		}
 		return false;
+	}
+
+	public void initialize() {
+		if (microprofileProjectListener != null) {
+			return;
+		}
+		microprofileProjectListener = new MicroProfileProjectListener();
+		MessageBusConnection connection = ApplicationManager.getApplication().getMessageBus().connect(project);
+		connection.subscribe(VirtualFileManager.VFS_CHANGES, microprofileProjectListener);
+		project.getMessageBus().connect(project).subscribe(ProjectTopics.MODULES, microprofileProjectListener);
 	}
 }
