@@ -26,6 +26,7 @@ import com.redhat.devtools.intellij.qute.psi.template.datamodel.SearchContext;
 import com.redhat.devtools.intellij.qute.psi.utils.AnnotationUtils;
 import com.redhat.devtools.intellij.qute.psi.utils.PsiTypeUtils;
 import com.redhat.qute.commons.datamodel.resolvers.ValueResolverInfo;
+import com.redhat.qute.commons.datamodel.resolvers.ValueResolverKind;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.List;
@@ -103,7 +104,7 @@ public class TemplateDataAnnotationSupport extends AbstractAnnotationTypeReferen
 			return;
 		}
 
-		ITypeResolver typeResolver = QuteSupportForTemplate.createTypeResolver(type);
+		ITypeResolver typeResolver = QuteSupportForTemplate.createTypeResolver(type, context.getJavaProject());
 
 		// Loop for static fields
 		PsiField[] fields = type.getFields();
@@ -130,12 +131,13 @@ public class TemplateDataAnnotationSupport extends AbstractAnnotationTypeReferen
 				resolver.setSourceType(sourceType);
 				resolver.setSignature(typeResolver.resolveSignature(member));
 				resolver.setNamespace(StringUtils.isNotEmpty(namespace) ? namespace : sourceType.replace('.', '_'));
+				resolver.setKind(ValueResolverKind.TemplateData);
 				if (!resolvers.contains(resolver)) {
 					resolvers.add(resolver);
 				}
 			}
 		} catch (Exception e) {
-			LOGGER.log(Level.SEVERE,
+			LOGGER.log(Level.WARNING,
 					"Error while getting annotation member value of '" + member.getName() + "'.", e);
 		}
 	}
