@@ -54,8 +54,8 @@ public class GenerateOpenAPIAnnotationsTest extends MavenModuleImportingTestCase
 
 		String newText = "package org.acme.openapi;\n\nimport org.eclipse.microprofile.openapi.annotations.Operation;\n\nimport java.util.Properties;\nimport javax.enterprise.context.RequestScoped;\nimport javax.ws.rs.GET;\nimport javax.ws.rs.Path;\nimport javax.ws.rs.core.Response;\n\n@RequestScoped\n@Path(\"/systems\")\npublic class NoOperationAnnotation {\n\n    @Operation(summary = \"\", description = \"\")\n    @GET\n    public Response getMyInformation(String hostname) {\n        return Response.ok(listContents()).build();\n    }\n\n    @Operation(summary = \"\", description = \"\")\n    @GET\n    public Response getPropertiesForMyHost() {\n        return Response.ok().build();\n    }\n\n    @Operation(summary = \"\", description = \"\")\n    private Properties listContents() {\n        Properties info = new Properties();\n        info.setProperty(\"Name\", \"APITest\");\n        info.setProperty(\"Desc\", \"API Test\");\n        return info;\n    }\n\n}";
 		assertJavaCodeAction(codeActionParams, utils,
-				ca(uri, "Generate OpenAPI Annotations", d,
-						te(0, 0, 29, 1, newText))
+				ca(uri, "Generate OpenAPI Annotations for 'NoOperationAnnotation'", d,
+						te(0, 0, 34, 1, newText))
 		);
 	}
 
