@@ -76,6 +76,85 @@ public class MicroProfileRestClientJavaCodeLensTest extends MavenModuleImporting
 		assertCodeLenses("https://restcountries.uri/rest", params, utils);
 	}
 
+	@Test
+	public void testUrlCodeLensPropertiesWithPropertyExpression() throws Exception {
+		Module javaProject = createMavenModule(new File("projects/lsp4mp/projects/maven/rest-client-quickstart"));
+		IPsiUtils utils = PsiUtilsLSImpl.getInstance(myProject);
+
+		// Initialize file
+		initConfigFile(javaProject);
+
+		MicroProfileJavaCodeLensParams params = new MicroProfileJavaCodeLensParams();
+		VirtualFile javaFile = LocalFileSystem.getInstance().refreshAndFindFileByPath(ModuleUtilCore.getModuleDirPath(javaProject) + "/src/main/java/org/acme/restclient/CountriesService.java");
+		String uri = VfsUtilCore.virtualToIoFile(javaFile).toURI().toString();
+		params.setUri(uri);
+		params.setUrlCodeLensEnabled(true);
+
+		// No configuration of base url
+		List<? extends CodeLens> lenses = PropertiesManagerForJava.getInstance().codeLens(params, utils);
+		Assert.assertEquals(0, lenses.size());
+
+		// /mp-rest/url
+		// get final value by following expressions transitively
+		saveFile(MicroProfileConfigSourceProvider.MICROPROFILE_CONFIG_PROPERTIES_FILE,
+				"org.acme.restclient.CountriesService/mp-rest/url = https://restcountries.url/${asdf}\nasdf=${hjkl}\nhjkl=rest",
+				javaProject);
+		assertCodeLenses("https://restcountries.url/rest", params, utils);
+
+		// /mp-rest/uri
+		// get final value though using default
+		saveFile(MicroProfileConfigSourceProvider.MICROPROFILE_CONFIG_PROPERTIES_FILE,
+				"org.acme.restclient.CountriesService/mp-rest/url = https://restcountries.url/${ASDF:rest}",
+				javaProject);
+		assertCodeLenses("https://restcountries.url/rest", params, utils);
+
+		// /mp-rest/uri
+		// cyclic dependencies between properties
+		// this will likely prevent the micro profile application from starting,
+		// but we also want to prevent the language server from freezing
+		saveFile(MicroProfileConfigSourceProvider.MICROPROFILE_CONFIG_PROPERTIES_FILE,
+				"org.acme.restclient.CountriesService/mp-rest/url = https://restcountries.url/${asdf}\nasdf=${hjkl}\nhjkl=${org.acme.restclient.CountriesService/mp-rest/url}",
+				javaProject);
+		assertCodeLenses("https://restcountries.url/${asdf}", params, utils);
+
+		// /mp-rest/uri
+		// self loop
+		// this will also prevent the micro profile application from starting
+		saveFile(MicroProfileConfigSourceProvider.MICROPROFILE_CONFIG_PROPERTIES_FILE,
+				"org.acme.restclient.CountriesService/mp-rest/url = https://restcountries.url/${org.acme.restclient.CountriesService/mp-rest/url}",
+				javaProject);
+		assertCodeLenses("https://restcountries.url/${org.acme.restclient.CountriesService/mp-rest/url}", params,
+				utils);
+
+		// /mp-rest/uri
+		// double reference
+		saveFile(MicroProfileConfigSourceProvider.MICROPROFILE_CONFIG_PROPERTIES_FILE,
+				"org.acme.restclient.CountriesService/mp-rest/url = https://restcountries.url/${asdf}${asdf}\nasdf=rest",
+				javaProject);
+		assertCodeLenses("https://restcountries.url/restrest", params,
+				utils);
+
+		// /mp-rest/uri
+		// billion laughs
+		saveFile(MicroProfileConfigSourceProvider.MICROPROFILE_CONFIG_PROPERTIES_FILE,
+				"org.acme.restclient.CountriesService/mp-rest/url = https://restcountries.url/${lulz}\n" //
+						+ "lulz=${lol9}${lol9}${lol9}${lol9}${lol9}${lol9}${lol9}${lol9}${lol9}\n" //
+						+ "lol9=${lol8}${lol8}${lol8}${lol8}${lol8}${lol8}${lol8}${lol8}${lol8}\n" //
+						+ "lol8=${lol7}${lol7}${lol7}${lol7}${lol7}${lol7}${lol7}${lol7}${lol7}\n" //
+						+ "lol7=${lol6}${lol6}${lol6}${lol6}${lol6}${lol6}${lol6}${lol6}${lol6}\n" //
+						+ "lol6=${lol5}${lol5}${lol5}${lol5}${lol5}${lol5}${lol5}${lol5}${lol5}\n" //
+						+ "lol5=${lol4}${lol4}${lol4}${lol4}${lol4}${lol4}${lol4}${lol4}${lol4}\n" //
+						+ "lol4=${lol3}${lol3}${lol3}${lol3}${lol3}${lol3}${lol3}${lol3}${lol3}\n" //
+						+ "lol3=${lol2}${lol2}${lol2}${lol2}${lol2}${lol2}${lol2}${lol2}${lol2}\n" //
+						+ "lol2=${lol1}${lol1}${lol1}${lol1}${lol1}${lol1}${lol1}${lol1}${lol1}\n" //
+						+ "lol1=${lol}${lol}${lol}${lol}${lol}${lol}${lol}${lol}${lol}${lol}\n" //
+						+ "lol=lol",
+				javaProject);
+		assertCodeLenses("https://restcountries.url/${lulz}", params,
+				utils);
+
+	}
+
 	private static void initConfigFile(Module javaProject) throws IOException {
 		saveFile(MicroProfileConfigSourceProvider.MICROPROFILE_CONFIG_PROPERTIES_FILE, "", javaProject);
 		saveFile(QuarkusConfigSourceProvider.APPLICATION_PROPERTIES_FILE, "", javaProject);
@@ -154,7 +233,7 @@ public class MicroProfileRestClientJavaCodeLensTest extends MavenModuleImporting
 
 	private static void assertCodeLenses(String baseURL, MicroProfileJavaCodeLensParams params, IPsiUtils utils) {
 		List<? extends CodeLens> lenses = PropertiesManagerForJava.getInstance().codeLens(params, utils);
-		Assert.assertEquals(2, lenses.size());
+		Assert.assertEquals(8, lenses.size());
 
 		// @GET
 		// @Path("/name/{name}")
@@ -169,6 +248,48 @@ public class MicroProfileRestClientJavaCodeLensTest extends MavenModuleImporting
 		CodeLens lensForGetSingle = lenses.get(1);
 		Assert.assertNotNull(lensForGetSingle.getCommand());
 		Assert.assertEquals(baseURL + "/v2/name/{name}", lensForGetSingle.getCommand().getTitle());
+
+		// @POST
+		// @Path("/post")
+		// String post();
+		CodeLens lensForPost = lenses.get(2);
+		Assert.assertNotNull(lensForPost.getCommand());
+		Assert.assertEquals(baseURL + "/v2/post", lensForPost.getCommand().getTitle());
+
+		// @PUT
+		// @Path("/put")
+		// String put();
+		CodeLens lensForPut = lenses.get(3);
+		Assert.assertNotNull(lensForPut.getCommand());
+		Assert.assertEquals(baseURL + "/v2/put", lensForPut.getCommand().getTitle());
+
+		// @DELETE
+		// @Path("/delete")
+		// String delete();
+		CodeLens lensForDelete = lenses.get(4);
+		Assert.assertNotNull(lensForDelete.getCommand());
+		Assert.assertEquals(baseURL + "/v2/delete", lensForDelete.getCommand().getTitle());
+
+		// @HEAD
+		// @Path("/head")
+		// String head();
+		CodeLens lensForHead = lenses.get(5);
+		Assert.assertNotNull(lensForHead.getCommand());
+		Assert.assertEquals(baseURL + "/v2/head", lensForHead.getCommand().getTitle());
+
+		// @OPTIONS
+		// @Path("/options")
+		// String options();
+		CodeLens lensForOptions = lenses.get(6);
+		Assert.assertNotNull(lensForOptions.getCommand());
+		Assert.assertEquals(baseURL + "/v2/options", lensForOptions.getCommand().getTitle());
+
+		// @PATCH
+		// @Path("/patch")
+		// String patch();
+		CodeLens lensForPatch = lenses.get(7);
+		Assert.assertNotNull(lensForPatch.getCommand());
+		Assert.assertEquals(baseURL + "/v2/patch", lensForPatch.getCommand().getTitle());
 	}
 
 }
